Take all tests of LineInput in a separate file. Rename test_gfx_ui to test_gfx_ui_base
